If no processed trace is returned, don't pass to coverage (#85)
The use can override the process_trace function to sanitise or modify
the trace data before sending to coverage.

In some cases, this may mean that no coverage is required to be covered
for this trace packet.
Eg. Data was deemed irrelevant for coverage, or data is cached while
waiting for additional trace data.

If no trace data is required to be passed to coverage, process_trace can
return None. This will not be passed to coverage.